Only render results if exec'ed q matches current q
There was a potential race condition when a user typed a query quickly.
Imagine that a user quickly typed the query "RA", and the "RA" results
came back before the "R" results.  In this case, the "R" results
would've overwritten the "RA" results.

This commit fixes that potential race condition.