Work around out of order XHR responses
Prior to this commit, if a user typed quickly on an unreliable network,
this could've happened:

1. User searches for `wee`
2. Spool issues an XHR request for `wee`
3. User searches for `ween`
4. Spool issues an XHR request for `ween`
5. Results for `ween` come in and get rendered
6. Results for `wee` come in and a bunch of Weezer songs get rendered

Our approach is for the back-end to return the search metadata including
the search query.  Then for the front-end to only render results if the
query in the metadata matches the (last) query stored in Redux.