Clear search when Search component mounts
Previously, we experienced the following bug:

1. Issue a search but don't click a result
2. Click the logo to navigate to `/wtf`
3. Click the logo again to nagivate back to the video

Upon navigating back to the video, the search input remained populated
with your search query (but blurred, so small search bar) and the
results remained visible.

This commit fixes that bug.  Now, when you navigate back to the video,
the input is empty (with the `search` placeholder) and there are no
results.